+++ stan/math/prim/fun/square.hpp
@@ -1,9 +1,8 @@
 #ifndef STAN_MATH_PRIM_FUN_SQUARE_HPP
 #define STAN_MATH_PRIM_FUN_SQUARE_HPP
 
-#include <stan/math/prim/meta.hpp>
 #include <stan/math/prim/fun/Eigen.hpp>
-#include <stan/math/prim/fun/square.hpp>
+#include <stan/math/prim/meta.hpp>
 #include <stan/math/prim/functor/apply_scalar_unary.hpp>
 #include <stan/math/prim/functor/apply_vector_unary.hpp>
 #include <cmath>
@@ -49,7 +48,7 @@ struct square_fun {
  */
 template <
     typename Container,
-    require_not_container_st<std::is_arithmetic, Container>* = nullptr,
+    require_not_stan_scalar_t<Container>* = nullptr,
     require_not_var_matrix_t<Container>* = nullptr,
     require_not_nonscalar_prim_or_rev_kernel_expression_t<Container>* = nullptr>
 inline auto square(const Container& x) {
